refactor: use BaseElement as parent class for AlertElement and PageElement
