+++ page/alert_element.py
@@ -1,24 +1,19 @@
 from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
+from page.base_element import BaseElement
 
 
-class AlertElement:
+class AlertElement(BaseElement):
     def __init__(self, driver):
-        self.driver = driver
-        self.alert = self.alert_element()
+        super().__init__(driver)
 
-    def alert_element(self):
-        return WebDriverWait(self.driver, 10).until(ec.alert_is_present())
+    def web_element(self):
+        return BaseElement.web_element(ec.alert_is_present())
 
     def accept(self):
-        self.alert.accept()
+        self.activate_element()
+        self.element.accept()
 
     def dismiss(self):
-        self.alert.dismiss()
-
-    def send_keys(self, text):
-        self.alert.send_keys(text)
-
-    @property
-    def text(self):
-        return self.alert.text
+        self.activate_element()
+        self.element.dismiss()
