+++ page/base_element.py
@@ -1,15 +1,11 @@
 from selenium.webdriver.common.keys import Keys
-from selenium.webdriver.support import expected_conditions as ec
+# from selenium.webdriver.support import expected_conditions as ec
 from selenium.webdriver.support.wait import WebDriverWait
 
 
 class BaseElement:
-    def __init__(self, by, value, driver, name=''):
+    def __init__(self, driver):
         self.driver = driver
-        self.by = by
-        self.value = value
-        self.name = name
-        self.locator = (self.by, self.value)
         self.element = None
         self.active_element = False
 
@@ -23,9 +19,8 @@ class BaseElement:
             self.element = None
         self.active_element = False
 
-    def web_element(self):
-        return WebDriverWait(self.driver, 10)\
-            .until(ec.visibility_of_element_located(locator=self.locator))
+    def web_element(self, ec):
+        return WebDriverWait(self.driver, 10).until(ec)
 
     def click(self):
         self.activate_element()
@@ -47,10 +42,6 @@ class BaseElement:
         self.activate_element()
         self.element.send_keys(Keys.CONTROL + 'a')
         self.element.send_keys(Keys.DELETE)
-
-    def select_drop_down(self, index):
-        self.activate_element()
-        self.element.select_by_index(index)
 
     @property
     def text(self):
