+++ page/page_element.py
@@ -1,0 +1,20 @@
+from selenium.webdriver.common.keys import Keys
+from selenium.webdriver.support import expected_conditions as ec
+from selenium.webdriver.support.wait import WebDriverWait
+from page.base_element import BaseElement
+
+
+class PageElement(BaseElement):
+    def __init__(self, by, value, driver, name=''):
+        super().__init__(driver)
+        self.by = by
+        self.value = value
+        self.name = name
+        self.locator = (self.by, self.value)
+
+    def web_element(self):
+        return BaseElement.web_element(ec.visibility_of_element_located(locator=self.locator))
+
+    def select_drop_down(self, index):
+        self.activate_element()
+        self.element.select_by_index(index)
